Refactor ChangeEmailDialog to use separate components.
